pyproject.toml
[tool.poetry]
name = "ml-course"
version = "0.1.0"
description = "A structured learning path for Machine Learning"
authors = ["Arbi Kullakshi <arbikullakshi0@gmail.com>"]
license = "MIT"
package-mode = true
readme = "README.md"
packages = [
    { include = "src" },
    { include = "tools" }
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
scikit-learn = "^1.6.0"
matplotlib = "^3.10.0"
scipy = "^1.14.1"
tensorflow = "^2.15.0"
tensorflow-addons = "^0.23.0"
tensorflow-hub = "^0.16.1"
tensorflow-datasets = "^4.9.7"
tensorboard = "^2.18.0"
typing-extensions = "^4.12.2"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
seaborn = "^0.13.2"
tensorflowjs = "^4.22.0"
notebook = "^7.3.2"
tensor-annotations-tensorflow-stubs = "^2.0.3"
pyyaml = "^6.0.2"
keras = "^3.8.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.14.1"
pre-commit = "^4.0.1"
coverage = "^7.6.10"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
jupyter-black = "^0.4.0"
types-pyyaml = "^6.0.12.20241230"
pydantic = { extras = ["mypy"], version = "^2.10.4" }
types-tqdm = "^4.67.0.20241221"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ml = "src.main:main"
lint = "tools.lint:run_lint"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/.venv
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100
src_paths = ["src", "tools", "tests"]
skip = [".venv"]
